+++ app/handler/project.go
@@ -1,16 +1,17 @@
-package app
+package handler
 
 import (
 	"context"
 	"time"
 
+	"github.com/minguu42/mtasks/app"
 	"github.com/minguu42/mtasks/app/logging"
 	"github.com/minguu42/mtasks/app/ogen"
 )
 
 // CreateProject は POST /projects に対応するハンドラ
 func (h *Handler) CreateProject(ctx context.Context, req *ogen.CreateProjectReq) (*ogen.Project, error) {
-	u, ok := ctx.Value(userKey{}).(*User)
+	u, ok := ctx.Value(userKey{}).(*app.User)
 	if !ok {
 		logging.Errorf("ctx.Value(userKey{}).(*User) failed")
 		return nil, errUnauthorized
@@ -27,7 +28,7 @@ func (h *Handler) CreateProject(ctx context.Context, req *ogen.CreateProjectReq)
 
 // ListProjects は GET /projects に対応するハンドラ
 func (h *Handler) ListProjects(ctx context.Context, params ogen.ListProjectsParams) (*ogen.Projects, error) {
-	u, ok := ctx.Value(userKey{}).(*User)
+	u, ok := ctx.Value(userKey{}).(*app.User)
 	if !ok {
 		logging.Errorf("ctx.Value(userKey{}).(*User) failed")
 		return nil, errUnauthorized
@@ -53,7 +54,7 @@ func (h *Handler) ListProjects(ctx context.Context, params ogen.ListProjectsPara
 
 // UpdateProject は PATCH /projects/{projectID} に対応するハンドラ
 func (h *Handler) UpdateProject(ctx context.Context, req *ogen.UpdateProjectReq, params ogen.UpdateProjectParams) (*ogen.Project, error) {
-	u, ok := ctx.Value(userKey{}).(*User)
+	u, ok := ctx.Value(userKey{}).(*app.User)
 	if !ok {
 		logging.Errorf("ctx.Value(userKey{}).(*User) failed")
 		return nil, errUnauthorized
@@ -87,7 +88,7 @@ func (h *Handler) UpdateProject(ctx context.Context, req *ogen.UpdateProjectReq,
 
 // DeleteProject は DELETE /projects/{projectID} に対応するハンドラ
 func (h *Handler) DeleteProject(ctx context.Context, params ogen.DeleteProjectParams) error {
-	u, ok := ctx.Value(userKey{}).(*User)
+	u, ok := ctx.Value(userKey{}).(*app.User)
 	if !ok {
 		logging.Errorf("ctx.Value(userKey{}).(*User) failed")
 		return errUnauthorized
