+++ app/handler/task.go
@@ -1,16 +1,17 @@
-package app
+package handler
 
 import (
 	"context"
 	"time"
 
+	"github.com/minguu42/mtasks/app"
 	"github.com/minguu42/mtasks/app/logging"
 	"github.com/minguu42/mtasks/app/ogen"
 )
 
 // CreateTask は POST /projects/{projectID}/tasks に対応するハンドラ
 func (h *Handler) CreateTask(ctx context.Context, req *ogen.CreateTaskReq, params ogen.CreateTaskParams) (*ogen.Task, error) {
-	u, ok := ctx.Value(userKey{}).(*User)
+	u, ok := ctx.Value(userKey{}).(*app.User)
 	if !ok {
 		logging.Errorf("ctx.Value(userKey{}).(*User) failed")
 		return nil, errUnauthorized
@@ -37,7 +38,7 @@ func (h *Handler) CreateTask(ctx context.Context, req *ogen.CreateTaskReq, param
 
 // ListTasks は GET /projects/{projectID}/tasks に対応するハンドラ
 func (h *Handler) ListTasks(ctx context.Context, params ogen.ListTasksParams) (*ogen.Tasks, error) {
-	u, ok := ctx.Value(userKey{}).(*User)
+	u, ok := ctx.Value(userKey{}).(*app.User)
 	if !ok {
 		logging.Errorf("ctx.Value(userKey{}).(*User) failed")
 		return nil, errUnauthorized
@@ -73,7 +74,7 @@ func (h *Handler) ListTasks(ctx context.Context, params ogen.ListTasksParams) (*
 
 // UpdateTask は PATCH /projects/{projectID}/tasks/{taskID} に対応するハンドラ
 func (h *Handler) UpdateTask(ctx context.Context, req *ogen.UpdateTaskReq, params ogen.UpdateTaskParams) (*ogen.Task, error) {
-	u, ok := ctx.Value(userKey{}).(*User)
+	u, ok := ctx.Value(userKey{}).(*app.User)
 	if !ok {
 		logging.Errorf("ctx.Value(userKey{}).(*User) failed")
 		return nil, errUnauthorized
@@ -119,7 +120,7 @@ func (h *Handler) UpdateTask(ctx context.Context, req *ogen.UpdateTaskReq, param
 
 // DeleteTask は DELETE /projects/{projectID}/tasks/{taskID} に対応するハンドラ
 func (h *Handler) DeleteTask(ctx context.Context, params ogen.DeleteTaskParams) error {
-	u, ok := ctx.Value(userKey{}).(*User)
+	u, ok := ctx.Value(userKey{}).(*app.User)
 	if !ok {
 		logging.Errorf("ctx.Value(userKey{}).(*User) failed")
 		return errUnauthorized
